requirements.txt
Pillow
pynput